Architecture Change. Component Card Faces.
The code base has changed substantially to accomodate for custom card faces in the future. You can now create a custom card view that can be generated by your own card which contain their own set of values and semantics.

For example, a suited card will need a Suit and a Value. A picture card may just need a value and maybe a scene. Each of these cards will have different ways of displaying their data and therefore need some kind of decorator.

The class CardFaceView can take a card as a parameter and return the View needed for that card. To add your own card type, simply add it to the methods in that class and create your own Card model than implements the <Card> protocol.

All classes are documented so if you seek further help, consult the documentation.

Added Stacks and Decks where decks are an extension of a Stack with slightly different behaviour.

Added Player model and tests.